- New Release Build (2736) (for German and Russian version)
Bugfix: Tactical Interface is still locked after QuickLoad (by Zathras)
The valid range for Price_Modifier is (-1)-100.  -1 & 0 were added as options some time back to allow the game to dynamically adjust the Price_Modifier value based on game progress.
Translated new text for Multiplayer and HAM.
Updates to the HAM/100AP integration.
Added Min/Max Suppression Tolerance values (from Headrock).
Bugfix: Took away the ability to reload a weapon from the item description box while using an ammo crate.
Corretion to the 100AP/HAM integration.
Made an APBPConstant a little clearer.  Before it was labeled as part of the suppression system when in reality it's not.
Fixes/Improvements to Option screen (by Arynn)
Added the new strings from the multiplayer branch
Added the new strings from the multiplayer branch
Updated a few "wrong" default values to "correct" ones. In multiplayer we read the default ja2_options.ini. Problem was, that min. and max. CTH was set to 0, which resulted in zero CTH
Translated MP Text in _GermanText.cpp
A few minor adjustments
Disabled some of the charges associated with the Increased Aiming Costs option.
Corrected a AP display issue when you had to turn while a weapon was not in a ready position.
Found a bug in the russian text file that was preventing the creation of russian builds.

git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameData@2744 3b4a5df2-a311-0410-b5c6-a8a6f20db521